- compile under 5.13 - fix fails like s/sv_undef/PL_sv_undef/ ( thanks to Slaven Rezic ) - disable lots of warnings for msvc v10 compiler

+++ AbstractMenu.c
@@ -448,7 +448,7 @@ new_av(  PMenuItemReg m, int level)
 	    SV * sv;
             int shift = ( m-> flags. checked ? 1 : 0) + ( m-> flags. disabled ? 1 : 0);
 	    if ( shift > 0) { /* has flags */
-	       int len = strlen( m-> variable);
+	       int len = (int) strlen( m-> variable);
                char * name = allocs( len + shift);
 	       if ( name) {
 		  int slen = len + shift;
